18-  Refatorando Exclusão de Cozinhas para Driven Domain Service

+++ src/main/java/com/algaworks/algaworkscurso/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.algaworks.algaworkscurso.infrastructure.repository;
 
 import com.algaworks.algaworkscurso.domain.model.Cozinha;
 import com.algaworks.algaworkscurso.domain.repository.CozinhaRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -37,8 +38,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Transactional
     @Override
-    public void remover(Cozinha cozinha) {
-        cozinha = buscar(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+        if(cozinha == null){
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinha);
     }
 }
